Commander: split out failure injection class into its own file
